Extract CoupoListItem data class
