+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/coupons/CouponListItem.kt
@@ -1,0 +1,8 @@
+package com.woocommerce.android.ui.coupons
+
+data class CouponListItem(
+    val id: Long,
+    val code: String? = null,
+    val summary: String,
+    val isActive: Boolean
+)
